Full new reply interface
