+++ lib/auth.js
@@ -89,7 +89,7 @@ internals.Auth.prototype.test = async function (name, request, next) {
 
     const bind = strategy.methods;
     const realm = strategy.realm;
-    const response = await request.server._blah.execute(strategy.methods.authenticate, request, { bind, realm, auth: true });
+    const response = await request.server._replier.execute(strategy.methods.authenticate, request, { bind, realm, auth: true });
 
     if (!response.isAuth) {
         return next(response);
@@ -257,11 +257,7 @@ internals.Auth.payload = async function (request) {
 
     const bind = strategy.methods;
     const realm = strategy.realm;
-    const response = await request.server._blah.execute(strategy.methods.payload, request, { bind, realm });
-
-    if (response === request.server._blah.continue) {
-        return;
-    }
+    const response = await request.server._replier.execute(strategy.methods.payload, request, { bind, realm });
 
     if (response &&
         response.isBoom &&
@@ -292,11 +288,7 @@ internals.Auth.response = async function (request) {
 
     const bind = strategy.methods;
     const realm = strategy.realm;
-    const response = await request.server._blah.execute(strategy.methods.response, request, { bind, realm });
-
-    if (response !== request.server._blah.continue) {
-        return response;
-    }
+    return await request.server._replier.execute(strategy.methods.response, request, { bind, realm, continue: 'undefined' });
 };
 
 
@@ -340,7 +332,7 @@ internals.Authenticator = class {
 
             const bind = strategy.methods;
             const realm = strategy.realm;
-            const response = await request.server._blah.execute(strategy.methods.authenticate, request, { bind, realm, auth: true });
+            const response = await request.server._replier.execute(strategy.methods.authenticate, request, { bind, realm, auth: true });
             if (!response.isAuth) {
                 return this.validate(response, null, next);
             }
