+++ lib/handler.js
@@ -49,7 +49,7 @@ internals.handler = async function (request, method, pre) {
 
     const bind = request.route.settings.bind;
     const realm = request.route.realm;
-    const response = await request.server._blah.execute(method, request, { bind, realm, continue: false });
+    const response = await request.server._replier.execute(method, request, { bind, realm, continue: 'null' });
 
     // Handler
