+++ lib/plugin.js
@@ -274,7 +274,6 @@ internals.Plugin.prototype.decorate = function (type, property, method, options)
 
     if (type === 'reply') {
         this.root._replier.decorate(property, method);
-        this.root._blah.decorate(property, method);
         this.root.decorations.reply.push(property);
         return;
     }
