+++ lib/reply.js
@@ -2,6 +2,7 @@
 
 // Load modules
 
+const Boom = require('boom');
 const Hoek = require('hoek');
 const Response = require('./response');
 
@@ -14,78 +15,153 @@ const internals = {};
 exports = module.exports = internals.Reply = function () {
 
     this._decorations = null;
+
+    this.abandon = Symbol('abandon');
+    this.close = Symbol('close');
+    this.continue = Symbol('continue');
 };
 
 
 internals.Reply.prototype.decorate = function (property, method) {
 
     Hoek.assert(!this._decorations || !this._decorations[property], 'Reply interface decoration already defined:', property);
-    Hoek.assert(['request', 'response', 'close', 'state', 'unstate', 'redirect', 'continue'].indexOf(property) === -1, 'Cannot override built-in reply interface decoration:', property);
+    Hoek.assert(['abandon', 'authenticated', 'close', 'continue', 'entity', 'redirect', 'response', 'realm', 'request', 'state', 'unauthenticated', 'unstate'].indexOf(property) === -1, 'Cannot override built-in reply interface decoration:', property);
 
     this._decorations = this._decorations || {};
     this._decorations[property] = method;
 };
 
 
+internals.Reply.prototype.execute = async function (method, request, options) {
+
+    const reply = this._interface(request, options);
+    const bind = options.bind || null;
+
+    let response;
+    try {
+        response = await (options.args ? method.call(bind, request, reply, ...options.args) : method.call(bind, request, reply));
+    }
+    catch (err) {
+        response = err instanceof Error ? Boom.boomify(err) : Boom.badImplementation('Unhandled rejected promise', err);
+    }
+
+    // Process response
+
+    if (response === undefined) {
+        response = Boom.badImplementation('Method did not return a value, a promise, or throw an error');
+    }
+
+    if (options.continue &&
+        response === this.continue) {
+
+        if (options.continue === 'undefined') {
+            return;
+        }
+
+        if (options.continue === 'null') {
+            response = null;
+        }
+    }
+
+    if (options.auth &&
+        response instanceof internals.Auth) {
+
+        return response;
+    }
+
+    if (typeof response !== 'symbol') {
+        response = Response.wrap(response, request);
+        if (!response.isBoom) {
+            response = await response._prepare();
+        }
+    }
+
+    return response;
+};
+
+
 /*
     const handler = function (request, reply) {
 
-        reply(error, result, ignore);   -> error || result (continue)
-        reply(...).takeover();          -> ... (continue)
+        reply(result);                      -> result (continue)
+        reply(result).takeover();           -> result (continue)
+        reply.continue;                     -> null (continue)
 
-        reply.continue(ignore);         -> null (continue)
-    };
+        throw error;                        -> error (respond)
+        reply(error);                       -> error (respond)
 
-    const ext = function (request, reply) {
-
-        reply(error, result, ignore);   -> error || result (respond)
-        reply(...).takeover();          -> ... (respond)
-
-        reply.continue(ignore);         -> (continue)
+        <undefined>                         -> badImplementation (respond)
     };
 
     const pre = function (request, reply) {
 
-        reply(error);                   -> error (respond)  // failAction override
-        reply(null, result, ignore);    -> result (continue)
-        reply(...).takeover();          -> ... (respond)
+        reply(result);                      -> result (continue)
+        reply(result).takeover();           -> result (skip other pre and handler)
+        reply.continue;                     -> null (continue)
 
-        reply.continue(ignore);         -> null (continue)
+        throw error;                        -> error (respond)      // failAction override
+        reply(error);                       -> error (respond)
+
+        <undefined>                         -> badImplementation (respond)
+    };
+
+    const ext = function (request, reply) {
+
+        reply(result);                      -> result (continue) - only in onPostHandler and onPreResponse, otherwise: badImplementation (respond)
+        reply(result).takeover();           -> result (respond)
+        reply.continue;                     -> (continue)
+
+        throw error;                        -> error (respond)
+        reply(error);                       -> error (respond)
+
+        <undefined>                         -> badImplementation (respond)
     };
 
     const auth = function (request, reply) {
 
-        reply(error, result, data);     -> error || result (respond) + data
-        reply(..., data).takeover();    -> ... (respond) + data
+        // Unauthenticated
 
-        reply.continue(data);           -> (continue) + data
+        reply(result);                      -> badImplementation (respond)
+        reply(result).takeover();           -> result (respond)
+        reply.continue;                     -> badImplementation (respond)
+
+        throw error;                        -> error (respond)
+        reply(error);                       -> error (respond)
+
+        <undefined>                         -> badImplementation (respond)
+
+        reply.unauthenticated(error, data)  -> error (respond) + data
+
+        // Authenticated
+
+        reply.authenticated(data )          -> (continue) + data
+
+        // Note: scheme.payload and scheme.response use the same interface as pre-handler extension methods
     };
 */
 
-internals.Reply.prototype.interface = function (request, realm, options, next) {        // next(err || response, data);
+internals.Reply.prototype._interface = function (request, options) {
 
-    const reply = (err, response, data) => {
+    const reply = (response) => reply.response(response);
 
-        Hoek.assert(data === undefined || options.data, 'Reply interface does not allow a third argument');
+    reply.abandon = this.abandon;
+    reply.close = this.close;
+    reply.continue = this.continue;
 
-        reply._data = data;                 // Held for later
-        return reply.response(err !== null && err !== undefined ? err : response);
-    };
-
-    reply.continue = internals.continue;
+    reply.entity = internals.entity;
     reply.redirect = internals.redirect;
     reply.response = internals.response;
-    reply.entity = internals.entity;
 
-    reply._settings = options;
-    reply._replied = false;
-    reply._next = Hoek.once(next);
-
-    reply.realm = realm;
+    reply.realm = options.realm;
     reply.request = request;
 
     reply.state = internals.state;
     reply.unstate = internals.unstate;
+
+    if (options.auth) {
+        reply.authenticated = internals.authenticated;
+        reply.unauthenticated = internals.unauthenticated;
+    }
 
     if (this._decorations) {
         const methods = Object.keys(this._decorations);
@@ -99,66 +175,15 @@ internals.Reply.prototype.interface = function (request, realm, options, next) {
 };
 
 
-internals.continue = function (data) {
+internals.response = function (result) {
 
-    if (data !== undefined) {
-        if (this._settings.data) {
-            this._data = data;
-        }
-        else if (this._settings.postHandler) {
-            const next = this._next;
-            this._next = (result) => next(null, result);            // Shift argument to signal not to stop processing other extensions
-            return this.response(data);
-        }
-        else {
-            throw new Error('reply.continue() does not allow any arguments');
-        }
-    }
-
-    this._next(null);
-    this._next = null;
-};
-
-
-internals.state = function (name, value, options) {
-
-    this.request._setState(name, value, options);
-};
-
-
-internals.unstate = function (name, options) {
-
-    this.request._clearState(name, options);
+    return Response.wrap(result, this.request);
 };
 
 
 internals.redirect = function (location) {
 
     return this.response('').redirect(location);
-};
-
-
-internals.response = function (result) {
-
-    Hoek.assert(!this._replied, 'reply interface called twice');
-    this._replied = true;
-
-    const response = Response.wrap(result, this.request);
-    if (response.isBoom) {
-        this._next(response);
-        this._next = null;
-        return response;
-    }
-
-    process.nextTick(() => {
-
-        if (this._next) {
-            response._prepare(this._next);
-            this._next = null;
-        }
-    });
-
-    return response;
 };
 
 
@@ -175,4 +200,40 @@ internals.entity = function (options) {
     }
 
     return null;
+};
+
+
+internals.state = function (name, value, options) {
+
+    this.request._setState(name, value, options);
+};
+
+
+internals.unstate = function (name, options) {
+
+    this.request._clearState(name, options);
+};
+
+
+internals.authenticated = function (data) {
+
+    Hoek.assert(data && data.credentials, 'Authentication data missing credentials information');
+
+    return new internals.Auth(null, data);
+};
+
+
+internals.unauthenticated = function (error, data) {
+
+    Hoek.assert(!data || data.credentials, 'Authentication data missing credentials information');
+
+    return new internals.Auth(error, data);
+};
+
+
+internals.Auth = function (error, data) {
+
+    this.isAuth = true;
+    this.error = error;
+    this.data = data;
 };
