+++ lib/request.js
@@ -404,7 +404,9 @@ internals.Request.prototype._execute = async function () {
             response = err;
         }
 
-        if (response) {
+        if (response &&
+            response !== this.server._replier.continue) {
+
             await this._reply(response);
             return;
         }
@@ -446,12 +448,12 @@ internals.Request.prototype._invoke = async function (event) {
         const ext = event.nodes[i];
         const bind = (ext.bind || ext.plugin.realm.settings.bind);
         const realm = ext.plugin.realm;
-        const response = await this.server._blah.execute(ext.func, this, { bind, realm });
+        const response = await this.server._replier.execute(ext.func, this, { bind, realm });
 
         // Process response
 
         if (typeof response === 'symbol') {
-            if (response === this.server._blah.continue) {
+            if (response === this.server._replier.continue) {
                 continue;
             }
 
@@ -496,12 +498,12 @@ internals.Request.prototype._reply = async function (exit) {
         this._setResponse(Response.wrap(exit, this));               // Wrap to ensure any object thrown is always a valid Boom or Response object
     }
 
-    if (this.response === this.server._blah.close) {
+    if (this.response === this.server._replier.close) {
         this.raw.res.end();                                     // End the response in case it wasn't already closed
     }
 
-    if (this.response === this.server._blah.abandon ||
-        this.response === this.server._blah.close) {
+    if (this.response === this.server._replier.abandon ||
+        this.response === this.server._replier.close) {
 
         this._finalize();
         return;
