+++ lib/route.js
@@ -371,7 +371,7 @@ internals.payload = async function (request) {
 
         // Custom handler
 
-        return await request.server._blah.execute(failAction, request, { realm: request.route.realm, args: [err] });
+        return await request.server._replier.execute(failAction, request, { realm: request.route.realm, args: [err] });
     }
 };
