+++ lib/server.js
@@ -19,7 +19,6 @@ const Podium = require('podium');
 const Statehood = require('statehood');
 
 const Auth = require('./auth');
-const Blah = require('./blah');
 const Compression = require('./compression');
 const Cors = require('./cors');
 const Defaults = require('./defaults');
@@ -72,7 +71,6 @@ exports = module.exports = internals.Server = function (options) {
     this._mime = new Mimos(this._settings.mime);
     this._registrations = {};                                                       // Tracks plugin for dependency validation { name -> { version } }
     this._replier = new Reply();
-    this._blah = new Blah();
     this._requestor = new Request();
     this._plugins = {};                                                             // Exposed plugin properties by name
     this._app = {};
