+++ lib/validation.js
@@ -114,8 +114,7 @@ internals.input = function (source, request, next) {
 
         // Custom handler
 
-        const reply = request.server._replier.interface(request, request.route.realm, {}, next);
-        request.route.settings.validate.failAction(request, reply, source, error);
+        request.server._replier.execute(request.route.settings.validate.failAction, request, { realm: request.route.realm, args: [source, error] }).then(next);
     };
 
     const localOptions = {
@@ -207,8 +206,7 @@ exports.response = function (request, next) {
 
         // Custom handler
 
-        const reply = request.server._replier.interface(request, request.route.realm, {}, next);
-        request.route.settings.response.failAction(request, reply, err);
+        request.server._replier.execute(request.route.settings.response.failAction, request, { realm: request.route.realm, args: [err] }).then(next);
     };
 
     const localOptions = {
